requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
tqdm>=4.65.0
pyyaml>=6.0

# Optional but recommended
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
